ready to trim project structures

+++ commons/src/claims.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Claims {
     pub nickname: String,
-    pub id:i32,
-    pub exp: usize,
+    pub id: i32,
+    pub exp: i64,
 }
